Added Problem 684 Added abs, floordiv, mod, powmod methods to BigInt class Added compatibility with int data type Added function for fibonacci numbers Fixed isless method in BigInt class Fixed importing and typing in Julia files

+++ problems/0006/compute.jl
@@ -1,1 +1,1 @@
-compute(n::Integer)::Integer = n * (n + 1) * (3 * n ^ 2 - n - 2) ÷ 12
+compute(n::Int64)::Int64 = n * (n + 1) * (3 * n ^ 2 - n - 2) ÷ 12
