+++ problems/0015/compute.jl
@@ -1,1 +1,1 @@
-compute(n::Integer)::Integer = binomial(2 * n, n)
+compute(n::Int64)::Int64 = binomial(2 * n, n)
