+++ problems/0015/solution.jl
@@ -1,7 +1,7 @@
 using BenchmarkTools
 BenchmarkTools.DEFAULT_PARAMETERS.samples = 100
 
-compute(n::Integer)::Integer = binomial(2 * n, n)
+compute(n::Int64)::Int64 = binomial(2 * n, n)
 
 compute(2)
